+++ src/lib.rs
@@ -258,6 +258,7 @@ pub mod fixed_codebook;
 pub mod gain_index_map;
 pub mod gain_predict;
 pub mod gain_reconstruct;
+pub mod long_term_postfilter;
 pub mod lp_synthesis;
 pub mod lsp_interpolate;
 pub mod lsp_reconstruct;